+++ ui_engine/bitmappery/src/components/cloud-file-selector/google-drive/google-drive-file-selector.vue
@@ -59,7 +59,7 @@ export default {
         this.retrieveFiles( pathToRetrieve );
     },
     methods: {
-        ...mapMutations([
+        ...mapMutations("bmp", [
             "setDriveConnected",
         ]),
         /* base component overrides */