+++ ui_engine/bitmappery/src/components/tool-options-panel/tool-options-text/tool-options-text.vue
@@ -132,7 +132,7 @@ export default {
         layerId: null,
     }),
     computed: {
-        ...mapGetters([
+        ...mapGetters("bmp", [
             "activeLayerIndex",
             "activeLayer",
         ]),
@@ -265,7 +265,7 @@ export default {
         this.handleBlur();
     },
     methods: {
-        ...mapMutations([
+        ...mapMutations("bmp", [
             "openDialog",
             "setActiveTool",
             "updateLayer",